BACKUP_CLEANUP_2025-12-24_113610/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
psycopg2-binary==2.9.9
stripe==14.1.0
python-dotenv==1.0.0
openai==1.54.0
httpx==0.27.0
PyMuPDF==1.26.6
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
reportlab==4.0.7
openpyxl==3.1.2
email-validator==2.1.0
sendgrid==6.11.0
gunicorn==21.2.0
numpy==1.26.2
pandas==2.1.4
sentence-transformers==2.5.0
scikit-learn==1.3.2
